09_streaming/pyproject.toml
[project]
name = "09-streaming"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "load-dotenv>=0.1.0",
    "openai-agents>=0.2.8",
]
